Combined two Utility classes intro one, general, Utility.java
